Create separate Preferences class (browser)
