+++ sdks/browser-sdk/src/Conversations.ts
@@ -2,7 +2,6 @@ import {
   ConversationType,
   type ConsentState,
   type Identifier,
-  type UserPreference,
 } from "@xmtp/wasm-bindings";
 import { v4 } from "uuid";
 import { AsyncStream, type StreamCallback } from "@/AsyncStream";
@@ -11,7 +10,6 @@ import { DecodedMessage } from "@/DecodedMessage";
 import { Dm } from "@/Dm";
 import { Group } from "@/Group";
 import type {
-  SafeConsent,
   SafeConversation,
   SafeCreateDmOptions,
   SafeCreateGroupOptions,
@@ -240,49 +238,5 @@ export class Conversations {
 
   async streamAllDmMessages(callback?: StreamCallback<DecodedMessage>) {
     return this.streamAllMessages(callback, ConversationType.Dm);
-  }
-
-  async streamConsent(callback?: StreamCallback<SafeConsent[]>) {
-    const streamId = v4();
-    const asyncStream = new AsyncStream<SafeConsent[]>();
-    const endStream = this.#client.handleStreamMessage<SafeConsent[]>(
-      streamId,
-      (error, value) => {
-        void asyncStream.callback(error, value ?? undefined);
-        void callback?.(error, value ?? undefined);
-      },
-    );
-    await this.#client.sendMessage("streamConsent", {
-      streamId,
-    });
-    asyncStream.onReturn = () => {
-      void this.#client.sendMessage("endStream", {
-        streamId,
-      });
-      endStream();
-    };
-    return asyncStream;
-  }
-
-  async streamPreferences(callback?: StreamCallback<UserPreference[]>) {
-    const streamId = v4();
-    const asyncStream = new AsyncStream<UserPreference[]>();
-    const endStream = this.#client.handleStreamMessage<UserPreference[]>(
-      streamId,
-      (error, value) => {
-        void asyncStream.callback(error, value ?? undefined);
-        void callback?.(error, value ?? undefined);
-      },
-    );
-    await this.#client.sendMessage("streamPreferences", {
-      streamId,
-    });
-    asyncStream.onReturn = () => {
-      void this.#client.sendMessage("endStream", {
-        streamId,
-      });
-      endStream();
-    };
-    return asyncStream;
   }
 }
