+++ sdks/browser-sdk/src/WorkerConversations.ts
@@ -1,13 +1,11 @@
 import {
   ConversationType,
-  type Consent,
   type ConsentState,
   type Conversation,
   type ConversationListItem,
   type Conversations,
   type Identifier,
   type Message,
-  type UserPreference,
 } from "@xmtp/wasm-bindings";
 import type { StreamCallback } from "@/AsyncStream";
 import {
@@ -177,28 +175,5 @@ export class WorkerConversations {
       { on_message, on_error },
       conversationType,
     );
-  }
-
-  streamConsent(callback?: StreamCallback<Consent[]>) {
-    const on_consent_update = (consent: Consent[]) => {
-      void callback?.(null, consent);
-    };
-    const on_error = (error: Error | null) => {
-      void callback?.(error, undefined);
-    };
-    return this.#conversations.streamConsent({ on_consent_update, on_error });
-  }
-
-  streamPreferences(callback?: StreamCallback<UserPreference[]>) {
-    const on_user_preference_update = (preferences: UserPreference[]) => {
-      void callback?.(null, preferences);
-    };
-    const on_error = (error: Error | null) => {
-      void callback?.(error, undefined);
-    };
-    return this.#conversations.streamPreferences({
-      on_user_preference_update,
-      on_error,
-    });
   }
 }
