+++ sdks/browser-sdk/test/Client.test.ts
@@ -1,4 +1,3 @@
-import { ConsentEntityType, ConsentState } from "@xmtp/wasm-bindings";
 import { v4 } from "uuid";
 import { describe, expect, it } from "vitest";
 import { Client } from "@/Client";
@@ -68,7 +67,7 @@ describe.concurrent("Client", () => {
     const user = createUser();
     const signer = createSigner(user);
     const client = await createRegisteredClient(signer);
-    const inboxState = await client.inboxState(false);
+    const inboxState = await client.preferences.inboxState(false);
     expect(inboxState.inboxId).toBe(client.inboxId);
     expect(inboxState.installations.map((install) => install.id)).toEqual([
       client.installationId,
@@ -81,7 +80,9 @@ describe.concurrent("Client", () => {
     const user2 = createUser();
     const signer2 = createSigner(user2);
     const client2 = await createClient(signer2);
-    const inboxState2 = await client2.getLatestInboxState(client.inboxId!);
+    const inboxState2 = await client2.preferences.getLatestInboxState(
+      client.inboxId!,
+    );
     expect(inboxState2.inboxId).toBe(client.inboxId);
     expect(inboxState.installations.length).toBe(1);
     expect(inboxState.installations[0].id).toBe(client.installationId);
@@ -106,7 +107,7 @@ describe.concurrent("Client", () => {
 
     await client.unsafe_addAccount(signer2);
 
-    const inboxState = await client.inboxState();
+    const inboxState = await client.preferences.inboxState();
     expect(inboxState.accountIdentifiers.length).toEqual(2);
     expect(inboxState.accountIdentifiers).toContainEqual(
       await signer.getIdentifier(),
@@ -127,7 +128,7 @@ describe.concurrent("Client", () => {
     await client.unsafe_addAccount(signer2);
     await client.removeAccount(await signer2.getIdentifier());
 
-    const inboxState = await client.inboxState();
+    const inboxState = await client.preferences.inboxState();
     expect(inboxState.accountIdentifiers).toEqual([
       await signer.getIdentifier(),
     ]);
@@ -145,7 +146,7 @@ describe.concurrent("Client", () => {
       dbPath: `./test-${v4()}.db3`,
     });
 
-    const inboxState = await client3.inboxState(true);
+    const inboxState = await client3.preferences.inboxState(true);
     expect(inboxState.installations.length).toBe(3);
 
     const installationIds = inboxState.installations.map((i) => i.id);
@@ -155,7 +156,7 @@ describe.concurrent("Client", () => {
 
     await client3.revokeAllOtherInstallations();
 
-    const inboxState2 = await client3.inboxState(true);
+    const inboxState2 = await client3.preferences.inboxState(true);
 
     expect(inboxState2.installations.length).toBe(1);
     expect(inboxState2.installations[0].id).toBe(client3.installationId);
@@ -172,7 +173,7 @@ describe.concurrent("Client", () => {
       dbPath: `./test-${v4()}.db3`,
     });
 
-    const inboxState = await client3.inboxState(true);
+    const inboxState = await client3.preferences.inboxState(true);
     expect(inboxState.installations.length).toBe(3);
 
     const installationIds = inboxState.installations.map((i) => i.id);
@@ -182,7 +183,7 @@ describe.concurrent("Client", () => {
 
     await client3.revokeInstallations([client.installationIdBytes!]);
 
-    const inboxState2 = await client3.inboxState(true);
+    const inboxState2 = await client3.preferences.inboxState(true);
 
     expect(inboxState2.installations.length).toBe(2);
 
@@ -190,44 +191,5 @@ describe.concurrent("Client", () => {
     expect(installationIds2).toContain(client2.installationId);
     expect(installationIds2).toContain(client3.installationId);
     expect(installationIds2).not.toContain(client.installationId);
-  });
-
-  it("should manage consent states", async () => {
-    const user1 = createUser();
-    const user2 = createUser();
-    const signer1 = createSigner(user1);
-    const signer2 = createSigner(user2);
-    const client1 = await createRegisteredClient(signer1);
-    const client2 = await createRegisteredClient(signer2);
-    const group = await client1.conversations.newGroup([client2.inboxId!]);
-
-    await client2.conversations.sync();
-    const group2 = await client2.conversations.getConversationById(group.id);
-
-    expect(group2).not.toBeNull();
-
-    expect(
-      await client2.getConsentState(ConsentEntityType.GroupId, group2!.id),
-    ).toBe(ConsentState.Unknown);
-
-    await client2.setConsentStates([
-      {
-        entityType: ConsentEntityType.GroupId,
-        entity: group2!.id,
-        state: ConsentState.Allowed,
-      },
-    ]);
-
-    expect(
-      await client2.getConsentState(ConsentEntityType.GroupId, group2!.id),
-    ).toBe(ConsentState.Allowed);
-
-    expect(await group2!.consentState()).toBe(ConsentState.Allowed);
-
-    await group2!.updateConsentState(ConsentState.Denied);
-
-    expect(
-      await client2.getConsentState(ConsentEntityType.GroupId, group2!.id),
-    ).toBe(ConsentState.Denied);
   });
 });
