+++ sdks/browser-sdk/test/Conversations.test.ts
@@ -1,15 +1,10 @@
-import {
-  ConsentEntityType,
-  ConsentState,
-  GroupPermissionsOptions,
-} from "@xmtp/wasm-bindings";
+import { ConsentState, GroupPermissionsOptions } from "@xmtp/wasm-bindings";
 import { v4 } from "uuid";
 import { describe, expect, it } from "vitest";
 import {
   createRegisteredClient,
   createSigner,
   createUser,
-  sleep,
 } from "@test/helpers";
 
 describe.concurrent("Conversations", () => {
@@ -492,7 +487,7 @@ describe.concurrent("Conversations", () => {
   });
 });
 
-describe("Conversations streaming", () => {
+describe("Streaming", () => {
   it("should stream new conversations", async () => {
     const user1 = createUser();
     const user2 = createUser();
@@ -770,117 +765,5 @@ describe("Conversations streaming", () => {
       }
     }
     expect(count).toBe(1);
-  });
-
-  it("should stream consent updates", async () => {
-    const user = createUser();
-    const user2 = createUser();
-    const signer1 = createSigner(user);
-    const signer2 = createSigner(user2);
-    const client = await createRegisteredClient(signer1);
-    const client2 = await createRegisteredClient(signer2);
-    const group = await client.conversations.newGroup([client2.inboxId!]);
-    const stream = await client.conversations.streamConsent();
-
-    await group.updateConsentState(ConsentState.Denied);
-
-    await sleep(1000);
-
-    await client.setConsentStates([
-      {
-        entity: group.id,
-        entityType: ConsentEntityType.GroupId,
-        state: ConsentState.Allowed,
-      },
-    ]);
-
-    await sleep(1000);
-    await client.setConsentStates([
-      {
-        entity: group.id,
-        entityType: ConsentEntityType.GroupId,
-        state: ConsentState.Denied,
-      },
-      {
-        entity: client2.inboxId!,
-        entityType: ConsentEntityType.InboxId,
-        state: ConsentState.Allowed,
-      },
-    ]);
-
-    setTimeout(() => {
-      void stream.callback(null, undefined);
-    }, 2000);
-
-    let count = 0;
-    for await (const updates of stream) {
-      if (updates === undefined) {
-        break;
-      }
-      count++;
-      if (count === 1) {
-        expect(updates.length).toBe(1);
-        expect(updates[0].entity).toBe(group.id);
-        expect(updates[0].entityType).toBe(ConsentEntityType.GroupId);
-        expect(updates[0].state).toBe(ConsentState.Denied);
-      }
-      if (count === 2) {
-        expect(updates.length).toBe(1);
-        expect(updates[0].entity).toBe(group.id);
-        expect(updates[0].entityType).toBe(ConsentEntityType.GroupId);
-        expect(updates[0].state).toBe(ConsentState.Allowed);
-      }
-      if (count === 3) {
-        expect(updates.length).toBe(2);
-        expect(updates[0].entity).toBe(group.id);
-        expect(updates[0].entityType).toBe(ConsentEntityType.GroupId);
-        expect(updates[0].state).toBe(ConsentState.Denied);
-        expect(updates[1].entity).toBe(client2.inboxId!);
-        expect(updates[1].entityType).toBe(ConsentEntityType.InboxId);
-        expect(updates[1].state).toBe(ConsentState.Allowed);
-      }
-    }
-    expect(count).toBe(3);
-  });
-
-  it("should stream preferences", async () => {
-    const user = createUser();
-    const signer = createSigner(user);
-    const client = await createRegisteredClient(signer);
-    const stream = await client.conversations.streamPreferences();
-
-    await sleep(2000);
-
-    const client2 = await createRegisteredClient(signer, {
-      dbPath: `./test-${v4()}.db3`,
-    });
-    const client3 = await createRegisteredClient(signer, {
-      dbPath: `./test-${v4()}.db3`,
-    });
-
-    await client3.conversations.syncAll();
-    await sleep(2000);
-    await client.conversations.syncAll();
-    await sleep(2000);
-    await client2.conversations.syncAll();
-    await sleep(2000);
-
-    setTimeout(() => {
-      void stream.callback(null, undefined);
-    }, 2000);
-
-    let count = 0;
-    for await (const preferences of stream) {
-      if (preferences === undefined) {
-        break;
-      }
-      count++;
-      expect(preferences).toBeDefined();
-      expect(preferences.length).toBe(1);
-      if (preferences[0].type === "HmacKeyUpdate") {
-        expect(preferences[0].key).toBeDefined();
-      }
-    }
-    expect(count).toBe(3);
   });
 });
